v2v: -o rhv-upload: Only set SSL context for https connections.
For real imageio servers the destination will always be https.  This
change has no effect there.

However when testing we want to use an http server for simplicity.  As
there is no certificate or cafile in this case the call to create the
context will fail.

This also simplifies creation of the context object and recognizes the
"insecure" flag for connecting to imageio.

Thanks: Nir Soffer.
(cherry picked from commit 6694028f9827a365385ba3ab199dc3938d8236df)